+++ src/config-file-watcher.ts
@@ -4,8 +4,8 @@ import * as vscodelc from 'vscode-languageclient/node';
 import {ClangdContext} from './clangd-context';
 import * as config from './config';
 
-export function activate(context: ClangdContext) {
-  if (config.get<string>('onConfigChanged') !== 'ignore') {
+export async function activate(context: ClangdContext) {
+  if (await config.get<string>('onConfigChanged') !== 'ignore') {
     context.client.registerFeature(new ConfigFileWatcherFeature(context));
   }
 }
@@ -19,20 +19,22 @@ class ConfigFileWatcherFeature implements vscodelc.StaticFeature {
   constructor(private context: ClangdContext) {}
   fillClientCapabilities(capabilities: vscodelc.ClientCapabilities) {}
 
-  initialize(capabilities: vscodelc.ServerCapabilities,
-             _documentSelector: vscodelc.DocumentSelector|undefined) {
-    if ((capabilities as ClangdClientCapabilities)
-            .compilationDatabase?.automaticReload)
+  async initialize(capabilities: vscodelc.ServerCapabilities,
+                   _documentSelector: vscodelc.DocumentSelector|undefined) {
+    if (!await config.get<boolean>('onConfigChangedForceEnable') &&
+        (capabilities as ClangdClientCapabilities)
+            .compilationDatabase?.automaticReload) {
       return;
+    }
     this.context.subscriptions.push(new ConfigFileWatcher(this.context));
   }
   getState(): vscodelc.FeatureState { return {kind: 'static'}; }
-  dispose() {}
+  clear() {}
 }
 
 class ConfigFileWatcher implements vscode.Disposable {
   private databaseWatcher?: vscode.FileSystemWatcher;
-  private debounceTimer?: NodeJS.Timer;
+  private debounceTimer?: NodeJS.Timeout;
 
   dispose() {
     if (this.databaseWatcher)
@@ -80,7 +82,7 @@ class ConfigFileWatcher implements vscode.Disposable {
     if ((await vscode.workspace.fs.stat(uri)).size <= 0)
       return;
 
-    switch (config.get<string>('onConfigChanged')) {
+    switch (await config.get<string>('onConfigChanged')) {
     case 'restart':
       vscode.commands.executeCommand('clangd.restart');
       break;