+++ src/extension.ts
@@ -1,34 +1,89 @@
 import * as vscode from 'vscode';
 
+import {ClangdExtension} from '../api/vscode-clangd';
+
+import {ClangdExtensionImpl} from './api';
 import {ClangdContext} from './clangd-context';
+import {get, update} from './config';
+
+let apiInstance: ClangdExtensionImpl|undefined;
 
 /**
  *  This method is called when the extension is activated. The extension is
  *  activated the very first time a command is executed.
  */
-export async function activate(context: vscode.ExtensionContext) {
+export async function activate(context: vscode.ExtensionContext):
+    Promise<ClangdExtension> {
   const outputChannel = vscode.window.createOutputChannel('clangd');
   context.subscriptions.push(outputChannel);
 
-  const clangdContext = new ClangdContext;
-  context.subscriptions.push(clangdContext);
+  let clangdContext: ClangdContext|null = null;
 
-  // An empty place holder for the activate command, otherwise we'll get an
-  // "command is not registered" error.
   context.subscriptions.push(
-      vscode.commands.registerCommand('clangd.activate', async () => {}));
+      vscode.commands.registerCommand('clangd.activate', async () => {
+        if (clangdContext && (clangdContext.clientIsStarting() ||
+                              clangdContext.clientIsRunning())) {
+          return;
+        }
+        vscode.commands.executeCommand('clangd.restart');
+      }));
   context.subscriptions.push(
       vscode.commands.registerCommand('clangd.restart', async () => {
-        await clangdContext.dispose();
-        await clangdContext.activate(context.globalStoragePath, outputChannel,
-                                     context.workspaceState);
+        if (!get<boolean>('enable')) {
+          vscode.window
+              .showInformationMessage(
+                  'Language features from Clangd are currently disabled. Would you like to enable them?',
+                  'Enable', 'Close')
+              .then(async (choice) => {
+                if (choice === 'Enable') {
+                  await update<boolean>('enable', true);
+                  vscode.commands.executeCommand('clangd.restart');
+                }
+              });
+          return;
+        }
+
+        // clangd.restart can be called when the extension is not yet activated.
+        // In such a case, vscode will activate the extension and then run this
+        // handler. Detect this situation and bail out (doing an extra
+        // stop/start cycle in this situation is pointless, and doesn't work
+        // anyways because the client can't be stop()-ped when it's still in the
+        // Starting state).
+        if (clangdContext && clangdContext.clientIsStarting()) {
+          return;
+        }
+        if (clangdContext)
+          clangdContext.dispose();
+        clangdContext = await ClangdContext.create(context.globalStoragePath,
+                                                   outputChannel);
+        if (clangdContext)
+          context.subscriptions.push(clangdContext);
+        if (apiInstance) {
+          apiInstance.client = clangdContext?.client;
+        }
+      }));
+  context.subscriptions.push(
+      vscode.commands.registerCommand('clangd.shutdown', async () => {
+        if (clangdContext && clangdContext.clientIsStarting()) {
+          return;
+        }
+        if (clangdContext)
+          clangdContext.dispose();
       }));
 
-  await clangdContext.activate(context.globalStoragePath, outputChannel,
-                               context.workspaceState);
+  let shouldCheck = false;
 
-  const shouldCheck = vscode.workspace.getConfiguration('clangd').get(
-      'detectExtensionConflicts');
+  if (vscode.workspace.getConfiguration('clangd').get<boolean>('enable')) {
+    clangdContext =
+        await ClangdContext.create(context.globalStoragePath, outputChannel);
+    if (clangdContext)
+      context.subscriptions.push(clangdContext);
+
+    shouldCheck = vscode.workspace.getConfiguration('clangd').get<boolean>(
+                      'detectExtensionConflicts') ??
+                  false;
+  }
+
   if (shouldCheck) {
     const interval = setInterval(function() {
       const cppTools = vscode.extensions.getExtension('ms-vscode.cpptools');
@@ -60,4 +115,7 @@ export async function activate(context: vscode.ExtensionContext) {
       }
     }, 5000);
   }
+
+  apiInstance = new ClangdExtensionImpl(clangdContext?.client);
+  return apiInstance;
 }